Refactor the autoload process in the bin script.
This is a bit easier to understand I think.

It also uses require rather than include, so that failure to properly
load the file terminates the script.

It also removes the error suppression which can silence unwanted errors
(It will silence any errors that occur while parsing the related class
(and any other files included in the process)).