Add options to exclude packages and dependencies from graph.
Dev dependencies of dependencies are not shown anymore, because they are not relevant for the current package.

* `--no-dev` will hide dev dependencies
* `--no-php` will hide the constraints regarding the PHP version
* `--no-ext` will hide PHP extensions
* `--depth` will limit the depth of the generated graph
* `--exclude-regex` allows to apply regular expressions to hide packages

To apply these filters the packages must be rendered in the correct order. Before the packages were drawn "randomly" and were connected later on. Now packages are drawn from root package and only if they are a dependency of a package. This allow filtering for dependencies and packages. If a package (node) is missing, then the dependency (edge) is not shown. Vice versa the dependency (edge) is not shown when the package (node) is not available.